+++ src/main/java/co/origamigames/quazimodded/QuaziModdedClient.java
@@ -14,6 +14,7 @@ public class QuaziModdedClient implements ClientModInitializer {
     // @SuppressWarnings({ "unchecked" })
     public void onInitializeClient() {
         // EntityRendererRegistry.INSTANCE.register(QMEntities.OWL, (dispatcher, context) -> new OwlEntityRenderer(dispatcher, new OwlEntityModel(), 0.7F));
+
         QMBlocks.setBlockRenderLayers();
     }
 }